@@ -116,46 +117,48 @@
     "webpack-merge": "^4.2.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.7.2",
-    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "@babel/preset-env": "^7.7.1",
-    "@babel/preset-flow": "^7.0.0",
-    "@babel/preset-react": "^7.7.0",
-    "@babel/preset-typescript": "^7.7.2",
-    "@babel/runtime": "^7.0.0",
-    "@commitlint/cli": "^8.2.0",
-    "@commitlint/config-conventional": "^8.2.0",
-    "@types/enzyme": "^3.10.3",
-    "@types/jest": "^24.0.23",
+    "@babel/core": "^7.8.3",
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
+    "@babel/preset-env": "^7.8.3",
+    "@babel/preset-flow": "^7.8.3",
+    "@babel/preset-react": "^7.8.3",
+    "@babel/preset-typescript": "^7.8.3",
+    "@babel/runtime": "^7.8.3",
+    "@commitlint/cli": "^8.3.4",
+    "@commitlint/config-conventional": "^8.3.4",
+    "@types/enzyme": "^3.10.4",
+    "@types/jest": "^24.9.0",
     "awesome-typescript-loader": "^5.2.1",
     "babel-core": "^7.0.0-bridge.0",
     "babel-jest": "^24.9.0",
     "babel-loader": "^8.0.6",
     "babel-plugin-add-module-exports": "^1.0.2",
-    "babel-plugin-module-resolver": "^3.2.0",
+    "babel-plugin-module-resolver": "^4.0.0",
     "babel-plugin-transform-object-assign": "^6.22.0",
     "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.24.1",
-    "chromedriver": "^78.0.1",
+    "chromedriver": "^79.0.0",
     "clean-webpack-plugin": "^3.0.0",
-    "copy-webpack-plugin": "^5.0.5",
-    "core-js": "^3.4.1",
-    "css-loader": "^3.2.0",
-    "cypress": "^3.6.1",
-    "enzyme": "^3.10.0",
-    "enzyme-adapter-react-16": "^1.15.1",
+    "copy-webpack-plugin": "^5.1.1",
+    "core-js": "^3.6.4",
+    "css-loader": "^3.4.2",
+    "cypress": "^3.8.2",
+    "enzyme": "^3.11.0",
+    "enzyme-adapter-react-16": "^1.15.2",
     "enzyme-to-json": "^3.4.3",
-    "eslint": "^6.6.0",
-    "file-loader": "^4.2.0",
+    "eslint": "^6.8.0",
+    "file-loader": "^5.0.2",
     "html-webpack-plugin": "^3.2.0",
-    "husky": "^3.0.9",
+    "husky": "^4.0.9",
     "identity-obj-proxy": "^3.0.0",
     "immutable": "^4.0.0-rc.12",
     "intl": "^1.2.5",
     "jest": "^24.9.0",
     "jest-resolve": "^24.9.0",
-    "lint-staged": "^9.4.3",
-    "mini-css-extract-plugin": "^0.8.0",
+    "jsdom": "^16.0.0",
+    "jsdom-global": "^3.0.2",
+    "lint-staged": "^9.5.0",
+    "mini-css-extract-plugin": "^0.9.0",
     "node-sass": "^4.13.0",
     "optimize-css-assets-webpack-plugin": "^5.0.3",
     "saas": "^1.0.0",
