@@ -1,3 +1,6 @@
+import AuthHeader from '@components/AuthHeader';
+import SingleTripCard from '@components/SingleTripCard';
+import TripCard from '@components/TripCard';
 import * as React from 'react';
 
 // third-party libraries
@@ -7,14 +10,14 @@ import * as moment from 'moment';
 import { connect } from 'react-redux';
 
 // pages
-import DashboardPage from '@pages/DashboardPage';
+import DashboardContainer from '@pages/DashboardContainer';
 
 // components
 import Button from '@components/Button';
 import Loader from '@components/Loader';
 
 // thunks
-import { getSingleTrip, requestTrip } from '@modules/trips';
+import { getAllUserTrips, getSingleTrip, requestTrip } from '@modules/trips';
 
 // styles
 import './SingleTripPage.scss';
@@ -23,132 +26,127 @@ import './SingleTripPage.scss';
 import TripsModal from '@components/TripsModal';
 import { SingleTripPageProps, SingleTripPageState } from './interfaces';
 
-export class SingleTripPage extends React.Component<SingleTripPageProps, SingleTripPageState> {
-  constructor(props) {
-    super(props);
-    this.state = {
-      isLoading: true,
-      modalOpen: false,
-    };
-  }
+export const SingleTripPage: React.FunctionComponent<SingleTripPageProps> = (props) => {
+  const [state, setState] = React.useState<SingleTripPageState>({
+    isLoading: true,
+    modalOpen: false,
+    trips: [],
+  });
 
-  handleModalOpen = () => {
-    this.setState({modalOpen: true});
-  };
-  handleModalClose = () => {
-    this.setState({modalOpen: false});
+  React.useEffect(() => {
+    const tripId = props.match.params.id;
+    props.getSingleTrip(tripId)
+      .then(() => setState({ ...state, isLoading: false }));
+  },              []);
+
+  React.useEffect(() => {
+    props.getAllUserTrips()
+      .then(() => setState({ ...state, trips: props.trips }))
+      .then(() => setState({ ...state, isLoading: false }));
+  },              []);
+
+  const handleModalOpen = () => {
+    setState({ ...state, modalOpen: true });
   };
 
-  handleRequestTrip = () => {
+  const handleModalClose = () => {
+    setState({ ...state, modalOpen: false });
+  };
+
+  const handleRequestTrip = () => {
     this.handleModalClose();
-    this.props.requestTrip(this.props.trip.id);
+    props.requestTrip(props.trip.id);
   };
 
-  componentDidMount() {
-    const tripId = this.props.match.params.id;
-    this.props.getSingleTrip(tripId)
-      .then(() => this.setState({ isLoading: false }));
-  }
-
-  renderHeaderContent = () => {
-    return (
-      <div className="cover cover-image">
-        <div className="head-title">
-        </div>
-      </div>
-    );
-  }
-
-  renderBodyContent = () => {
-    const { trip } = this.props;
+  const BodyContent = () => {
+    const { trip } = props;
 
     return (
       <div className="single-trip-page">
         <Grid>
           <Row>
             <Cell
-              className="mdc-layout-grid__cell grid-start-3
-                      mdc-layout-grid__cell--align-middle"
+              className="mdc-layout-grid__cell grid-start-3 mdc-layout-grid__cell--align-middle"
               columns={8}
               desktopColumns={8}
               tabletColumns={8}
               phoneColumns={4}
             >
               <div className="trip">
-                <div className="trip-row trip-row__rules">
-                  <div className="trip-row-content">
-                    <h4 className="trip-title">Overview</h4>
-                    <div className="trip-row-content__details">
-                      <div className="trip-content-places">
-                        <h3>{trip.origin}</h3>
-                          <MaterialIcon
-                            className="trip-flight-icon__origin"
-                            role="button"
-                            icon="flight"
-                            hasRipple={true}
-                            initRipple={null}/>
-                          <hr className="horizontal-line" />
-                          <MaterialIcon
-                            className="trip-flight-icon"
-                            role="button"
-                            icon="place"
-                            hasRipple={true}
-                            initRipple={null}/>
-                        <h3>{trip.destination}</h3>
-                      </div>
-                      <div className="trip-card">
-                        <div className="trip-card__details">
-                          <h5>Departure</h5>
-                          <h4>{`${moment(trip.departure_date).format('LL')}`}</h4>
-                        </div>
-                        <div className="trip-card__details">
-                          <h5>Arrival</h5>
-                          <h4>{`${moment(trip.arrival_date).format('LL')}`}</h4>
-                        </div>
-                        <div className="trip-card__details">
-                          <h5>Traveller</h5>
-                          <h4>
-                            {trip.traveller.first_name
-                              ? `${trip.traveller.first_name} ${trip.traveller.last_name}`
-                              : `${trip.traveller.username}`}
-                          </h4>
-                        </div>
-                      </div>
-                    </div>
+                <div className="trip-card">
+                  <div className="trip-card__details">
+                    <h3>{`Trip to ${trip.destination}`}</h3>
+                    <h4>
+                      {`${moment(trip.departure_date).format('LL')}`} -
+                      {`${moment(trip.arrival_date).format('LL')}`}
+                    </h4>
+                    <Cell
+                      className="mdc-layout-grid__cell grid-start-3
+                      mdc-layout-grid__cell--align-middle trip-button"
+                      align="middle"
+                    >
+                      <Button
+                        name="REQUEST"
+                        classes="mdc-button big-round-corner-button mdc-button--raised"
+                        onClick={handleModalOpen}
+                      />
+                      <TripsModal
+                        open={state.modalOpen}
+                        handleClose={handleModalClose}
+                        handleSubmitRequest={handleRequestTrip} />
+                    </Cell>
                   </div>
                 </div>
-
+                <div className="trip-content-places">
+                  <h3>{trip.origin}</h3>
+                  <MaterialIcon
+                    className="trip-flight-icon__origin"
+                    role="button"
+                    icon="flight"
+                    hasRipple={true}
+                    initRipple={null}/>
+                  <hr className="horizontal-line" />
+                  <MaterialIcon
+                    className="trip-flight-icon"
+                    role="button"
+                    icon="place"
+                    hasRipple={true}
+                    initRipple={null}/>
+                  <h3>{trip.destination}</h3>
+                </div>
               </div>
             </Cell>
           </Row>
           <Row>
-            <Cell
-              className="mdc-layout-grid__cell grid-start-3
-                      mdc-layout-grid__cell--align-middle trip-button"
-              align="middle"
-            >
-              <Button
-                name="Request"
-                classes="mdc-button big-round-corner-button mdc-button--raised"
-                onClick={this.handleModalOpen}
-              />
-              <TripsModal
-                open={this.state.modalOpen}
-                handleClose={this.handleModalClose}
-                handleSubmitRequest={this.handleRequestTrip} />
-            </Cell>
+            {
+              props.trips.map((trip) => {
+                return (
+                  <TripCard
+                    key={trip.id}
+                    trip={trip}
+                    redirect={this.redirectToSingleTrip}
+                    requestTrip={this.handleSubmitTripRequest}
+                  />
+                );
+              })
+            }
           </Row>
         </Grid>
       </div>
     );
-  }
+  };
 
-  renderTripsComponent = () => {
-    const { isLoading } = this.state;
-    return (
-      isLoading
+  return (
+    state.isLoading
       ? <Loader />
-      : <Grid>
+      :
+      <div className="register">
+      <AuthHeader
+        forwardButtonName={'Home'}
+        backwardButtonName={'Back'}
+        forwardLink={'/'}
+        backwardLink={'/explore'}/>
+      <Grid>
         <Row>
           <Cell
             columns={12}
@@ -156,29 +154,22 @@ export class SingleTripPage extends React.Component<SingleTripPageProps, SingleT
             tabletColumns={8}
             phoneColumns={4}
           >
-            {this.renderHeaderContent()}
-            {this.renderBodyContent()}
+            {BodyContent()}
           </Cell>
         </Row>
       </Grid>
-    );
-  }
-
-  render() {
-    return (
-     <React.Fragment>
-      <DashboardPage component={this.renderTripsComponent()}/>
-    </React.Fragment>
-    );
-  }
-}
+      </div>
+  );
+};
 
 export const mapStateToProps = state => ({
+  trips: state.trips.user_trips,
   trip: state.trips.trip,
   error: state.error,
 });
 
 export const mapDispatchToProps = dispatch => ({
+  getAllUserTrips: () => dispatch(getAllUserTrips()),
   getSingleTrip: tripId => dispatch(getSingleTrip(tripId)),
   requestTrip: tripIp => dispatch(requestTrip(tripIp)),
 });
