@@ -47,10 +47,10 @@ module.exports = {
   collectCoverageFrom: ['src/**/*.ts', 'src/**/*.tsx', '!src/**/interface.d.ts', '!src/**/*interfaces.d.ts'],
   coverageThreshold: {
     global: {
-      'branches': 20,
+      'branches': 10,
       'functions': 10,
-      'lines': 20,
-      'statements': 25
+      'lines': 10,
+      'statements': 15
     }
   },
   setupFiles: [
