@@ -1,7 +1,7 @@
-import { Trip } from 'modules/trips/interfaces';
+import { Trip } from '@modules/trips/interfaces';
 
 export interface TripCardProps {
   trip: Trip;
-  redirect: (tripId: string) => void;
-  requestTrip: (tripId: string) => void;
+  redirect?: (tripId: string) => void;
+  requestTrip?: (tripId: string) => void;
 }
