@@ -1,3 +1,5 @@
+import SearchInput from '@components/SearchInput';
+import { TextField } from '@material-ui/core';
 import * as React from 'react';
 
 // third-party libraries
@@ -7,7 +9,6 @@ import {
   Row
 } from '@material/react-layout-grid';
 import MaterialIcon from '@material/react-material-icon';
-import TextField, { Input } from '@material/react-text-field';
 import { connect } from 'react-redux';
 
 // components
@@ -19,7 +20,7 @@ import { displaySnackMessage } from '@modules/snack';
 import { getAllTrips, requestTrip } from '@modules/trips';
 
 // pages
-import DashboardPage from '@pages/DashboardPage';
+import DashboardContainer from '@pages/DashboardContainer';
 
 // interfaces
 import { ExplorePageProps, ExplorePageState } from './interfaces';
@@ -48,20 +49,7 @@ export class ExplorePage extends React.Component<ExplorePageProps, ExplorePageSt
 
   renderSearchField = () => (
     <div className="form-cell">
-      <TextField
-        className="mdc-text-field--fullwidth search"
-        outlined
-        label="Search for trips"
-        leadingIcon={<MaterialIcon role="button" icon="search" initRipple={null}/>}
-      >
-        <Input
-          className=""
-          value=""
-          name="search"
-          id="10"
-          type="text"
-        />
-      </TextField>
+      {SearchInput}
     </div>
   )
 
@@ -75,20 +63,6 @@ export class ExplorePage extends React.Component<ExplorePageProps, ExplorePageSt
       <Grid>
         <Row>
           <Cell
-            columns={12}
-            desktopColumns={12}
-            tabletColumns={8}
-            phoneColumns={4}
-          >
-            <div className="cover cover-image">
-              <div className="head-title">
-                <div className="title-cover title-cover-page">Explore</div>
-              </div>
-            </div>
-          </Cell>
-        </Row>
-        <Row>
-          <Cell
             className="mdc-layout-grid__cell grid-start-5 mdc-layout-grid__cell--align-middle"
             align="middle"
             order={5}
@@ -97,7 +71,7 @@ export class ExplorePage extends React.Component<ExplorePageProps, ExplorePageSt
             tabletColumns={4}
             phoneColumns={4}
           >
-            {this.renderSearchField()}
+              <SearchInput />
           </Cell>
         </Row>
         <Row>
@@ -126,7 +100,7 @@ export class ExplorePage extends React.Component<ExplorePageProps, ExplorePageSt
       isLoading
       ? <Loader />
       : <React.Fragment>
-        <DashboardPage component={this.renderExploreContent(trips)} />
+        <DashboardContainer component={this.renderExploreContent(trips)} />
       </React.Fragment>
     );
   }
